add example, don't check in binaries, the main entrypoint was somehow not performant so I'm trying this
