+++ examples/for-learning-and-testing/wasm/node-full/template/greeting.txt
@@ -1,0 +1,5 @@
+Hello, {{ .Name | upper }}!
+
+Items:
+{{ range .Items }}- {{ . }}
+{{ end }}
