+++ examples/for-learning-and-testing/wasm/node/demo.mjs
