+++ examples/for-learning-and-testing/wasm/node/mise.toml
